refactor: split the token into Light, base and Debt variants over a shared   library
