+++ Nargo.toml
@@ -1,11 +1,8 @@
-[package]
-name = "CMTAT"
-type = "contract"
-authors = [ "" ]
-compiler_version = ">=0.18.0"
-
-[dependencies]
-aztec = { git = "https://github.com/AztecProtocol/aztec-nr/", tag = "v5.2.0", directory = "aztec" }
-compressed_string = { git = "https://github.com/AztecProtocol/aztec-nr/", tag = "v5.2.0", directory = "compressed-string" }
-uint_note = { git = "https://github.com/AztecProtocol/aztec-nr/", tag = "v5.2.0", directory = "uint-note" }
-balance_set = { git = "https://github.com/AztecProtocol/aztec-nr/", tag = "v5.2.0", directory = "balance-set" }
+[workspace]
+members = [
+    "lib",
+    "contracts/cmtat-aztec",
+    "contracts/cmtat-aztec-debt",
+    "contracts/cmtat-aztec-light",
+]
+default-member = "contracts/cmtat-aztec"
