+++ contracts/cmtat-aztec-debt/src/test.nr
@@ -1,0 +1,4 @@
+mod utils;
+mod test_credit_events;
+mod test_debt_base;
+mod smoke;
