+++ contracts/cmtat-aztec-light/src/test.nr
@@ -1,0 +1,2 @@
+mod utils;
+mod smoke;
