+++ contracts/cmtat-aztec/src/test.nr
@@ -7,6 +7,4 @@ mod transfer_private;
 mod test_validation_module;
 mod test_enforcement_module;
 mod test_pause_module;
-mod test_credit_events;
-mod test_debt_base;
 mod test_extra_information;
