+++ contracts/cmtat-aztec/src/test/transfer_private.nr
@@ -1,5 +1,5 @@
 use crate::test::utils;
-use crate::CMTAToken as Token;
+use crate::CMTATAztec as Token;
 use aztec::test::helpers::authwit::add_private_authwit_from_call;
 
 global AUTHWIT_NONCE: Field = 7;
