+++ scripts/interaction.ts
@@ -2,7 +2,7 @@ import { AztecAddress } from "@aztec/aztec.js/addresses";
 import { createLogger } from "@aztec/aztec.js/log";
 import * as dotenv from 'dotenv';
 
-import { CMTATokenContract as TokenContract } from "../src/artifacts/CMTAToken.js";
+import { CMTATAztecContract as TokenContract } from "../src/artifacts/CMTATAztec.js";
 import { getAccountFromEnv } from "../src/utils/create_account_from_env.js";
 import { getSponsoredPaymentMethod } from "../src/utils/sponsored_fpc.js";
 import { setupWalletTestnet } from "../src/utils/setup_pxe_testnet.js";
