+++ src/test/e2e/accounts.test.ts
@@ -15,7 +15,7 @@ import type { AccountManager } from "@aztec/aztec.js/wallet";
 import type { EmbeddedWallet } from "@aztec/wallets/embedded";
 import { spawn } from 'child_process';
 
-import { CMTATokenContract as TokenContract } from "../../artifacts/CMTAToken.js";
+import { CMTATAztecContract as TokenContract } from "../../artifacts/CMTATAztec.js";
 import { getSponsoredPaymentMethod } from "../../utils/sponsored_fpc.js";
 import { setupWallet } from "../../utils/setup_pxe.js";
