--- src/types.nr
@@ -1,3 +1,0 @@
-// Superseded by the `balance_set` library (AztecProtocol/aztec-nr, directory "balance-set"), which
-// now ships the BalanceSet state variable this module used to hand-roll. Nothing references this
-// module any more - it is kept only because it could not be deleted in place, and should be removed.
