--- src/types/balance_set.nr
@@ -1,1 +1,0 @@
-// Superseded by `balance_set::BalanceSet` from AztecProtocol/aztec-nr. See ../types.nr.
